Hopeful fix for being logged out after logging in
